==> model_dev/3_EDA_and_data_preprocessing/31_data_preprocessing_formmating.ipynb <==
import pandas as pd

df = pd.read_csv("../../data/processed/model_input_dataset.csv") 
df.head()

### Formatting and Column Standardization

- Ensured all column names use lowercase with underscores (`snake_case`) for consistency.
- Converted categorical variables (e.g., `source`) to Pandas `category` dtype.
- Ensured the target column `label` is of integer type.
- No one-hot encoding was applied at this stage, as most ML models used (e.g., XGBoost, RF) handle categorical variables natively or do not require dummy variables.

# Step 1: Standardize Column Names
df.columns = df.columns.str.strip().str.lower().str.replace(" ", "_")
print("Standardized columns:", list(df.columns))

# Step 2: Check and Correct Data Types
print("Before type check:\n", df.dtypes)

# Convert LABEL to int
df['label'] = df['label'].astype(int)

print("\nAfter type check:\n", df.dtypes)


# Step 3: Save Cleaned Dataset
df.to_csv("../../data/processed/train_dataset_formatted.csv", index=False)
print("Saved cleaned dataset to 'train_dataset_formatted.csv'")
