==> model_dev/3_EDA_and_data_preprocessing/32_data_preprocessing_missing_value_and_duplicate.ipynb <==
import pandas as pd

# Load formatted dataset from previous step
df = pd.read_csv("../../data/processed/train_dataset_formatted.csv")
df.head()

## 1.Missing Values Summary
Radiometric features such as radio_k_pct, radio_th_ppm, radio_u_ppm, and derived ratios show ~3% missing values, Gravity feature, gravity_iso_residual show ~0.3% missing values. Other geophysical features are complete.

# Count missing values for each column
missing_summary = df.isnull().sum()
missing_summary = missing_summary[missing_summary > 0].sort_values(ascending=False)

print("Missing Values Summary:")
display(missing_summary)

# Show percentage for easier interpretation
missing_pct = (df.isnull().sum() / len(df)) * 100
print("Missing Value Percentage (%):")
display(missing_pct[missing_pct > 0].sort_values(ascending=False))

## 2. Missing Value Imputation Strategy

# List of features to impute (with missing values)
radiometric_features = [
    'radio_k_pct', 'radio_th_ppm', 'radio_u_ppm',
    'radio_th_k_ratio', 'radio_u_k_ratio', 'radio_u_th_ratio'
]

gravity_features = ['gravity_iso_residual', 'gravity_cscba']

magnetic_features = ['mag_uc_1_2km', 'mag_uc_2_4km', 'mag_uc_4_8km', 'mag_uc_8_12km', 'mag_uc_12_16km']

for col in radiometric_features + gravity_features + magnetic_features:
    df[col] = df.groupby('label')[col].transform(lambda x: x.fillna(x.median()))
print("Applied group-wise median imputation by source.")


# Recheck missing values after imputation
remaining_missing = df.isnull().sum()
print("Remaining missing values:")
display(remaining_missing[remaining_missing > 0])

## 3.Delete Duplicate Values

print(f"Duplicate rows before drop: {df.duplicated().sum()}")
df = df.drop_duplicates()
print(f"Duplicate rows after drop: {df.duplicated().sum()}")

df.to_csv("../../data/processed/train_dataset_formatted_no_missing.csv", index=False)
print("Cleaned dataset saved to 'train_dataset_formatted_no_missing.csv'")

## Missing Value Handling Summary

- **Columns affected**: Radiometric element concentrations and ratios, gravity_iso_residual, and magnetic features.
- **Imputation method used**: Median imputation.
  - Chosen because these features are highly skewed and contain physical/geochemical measurement values.
  - Median is robust to outliers and preserves distributional integrity.
- **Validation**: All missing values resolved after imputation.
- **Domain Justification**: Radiometric values are spatially interpolated from field measurements. Median preserves range while avoiding overfitting to extreme anomalies.
